Fixed KeyError if server is down
closes #15

Added an error check to get_map_images that confirms that a server dict
has a map key. This prevents a 500 Internal Error that would occur if a
server is down or nonresponsive.

I've also adjusted an if tag in index.html to skip a server box if the
server dict has an error rather than showing "server is down!" in a
poorly formatted box